DP-827: Adds the `watch` CLI command (#4)
Adds a new `watch` CLI command that watches a Gingr directory for new zipfiles and automatically processes them. New files should be placed in a `./ready` subdirectory. Successfully processed ones are moved to `./processed`, while ones with errors are moved to `./failed` along with an error log having the same name as the original file but a `.log` extension.